+++ Commands/radio.js
@@ -1,15 +1,24 @@
 const m3u8stream = require('m3u8stream');
 const https = require('https');
+const Channel = require('./Channel');
+const Logger = require('./Logger');
 /**
  * @usage !radio <optional: stop> <optional: master.m3u8-Link>
  * @does Plays Internetradio in your Channel
  */
+const defaultPlaylist = 'https://1livediggiuni-lh.akamaihd.net/i/1LIVEdiggi_HDS@179579/index_3_a-p.m3u8';
+var voiceChannel;
 module.exports = {
    radio: function (message) {
 
       let contentArgs = message.content.split(" "); //Split Message for simpler Access
 
-      const voiceChannel = message.member.voiceChannel;
+      voiceChannel = message.member.voice.channel;
+
+      if (voiceChannel == null) {
+         message.channel.send("Please join a VoiceChannel");
+         return;
+      }
 
       if (contentArgs[1] == 'stop') { //Stop playing
          try {
@@ -18,7 +27,19 @@ module.exports = {
          return;
       }
 
-      let playlist = 'https://1livediggiuni-lh.akamaihd.net/i/1LIVEdiggi_HDS@179579/index_3_a-p.m3u8'; //Default
+      let inChannel = false;
+      global.bot.voice.connections.every(conn => {
+         if (conn.channel.guild.id == message.guild.id) {
+            inChannel = true;
+         }
+      });
+
+      if (inChannel) {
+         message.channel.send("Im already in a Channel");
+         return;
+      }
+
+      let playlist = defaultPlaylist;//Default
 
       if (contentArgs[1]) {
          let isMaster = contentArgs[1].match(/(.*\/master\.m3u8).*/);
@@ -52,10 +73,14 @@ module.exports = {
                chunkReadahead: 20,
                highWaterMark: 1 << 25
             });
-            const dispatcher = connnection.playStream(stream);
+            const dispatcher = connnection.play(stream);
             dispatcher.on('end', () => {
                voiceChannel.leave();
             });
+         }).catch(ex => {
+            message.channel.send("Please join a Channel to use this Feature");
+            Logger.log(ex);
+            return;
          });
    }
 }